+++ CHANGELOG.md
@@ -14,3 +14,4 @@
 - PR #83 Fix conda upload
 - PR #84 Updated Docs and Readme with conda nightly install instructions for cuxfilter version 0.12
 - PR #86 Implemented #79 - cudatashader replaced by datashader(>=0.9) with cudf & dask_cudf support
+- PR #87 Implemented jupyter-server-proxy as discussed in #73